src/producers/requirements.txt
kafka-python
requests
websocket-client